pyproject.toml
[project]
name = "sandol_kakao_bot_service"
version = "0.1.0"
description = ""
authors = [{ name = "Seokyoung_Hong", email = "tjrdud0330@naver.com" }]
requires-python = ">=3.11,<3.12"
readme = "README.md"
dependencies = [
    "fastapi~=0.115.6",
    "uvicorn>=0.33,<0.34",
    "SQLAlchemy>=2,<3",
    "httpx>=0.28,<0.29",
    "fastapi-pagination>=0.12.34,<0.13",
    "python-dotenv>=1.0.1,<2",
    "aiosqlite>=0.21.0,<0.22",
    "alembic>=1.15.1,<2",
    "cachetools>=5.5.2,<6",
    "kakao-chatbot>=0.4.3,<0.5",
    "pytz>=2025.2",
    "diskcache>=5.6.3",
    "python-keycloak>=5.8.1",
    "cryptography>=46.0.3",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9,<0.10",
    "black>=25.1.0,<26",
    "mypy>=1.15.0,<2",
    "types-cachetools>=5.5.0.20240820,<6",
    "types-pytz>=2025.1.0.20250204,<2026",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "docs",
    "alembic"
]

line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C", "D", "R", "B", "N", "S", "PL"]
ignore = ["E203", "E501", "D415", "D403"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_|__)$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.pydocstyle]
convention = "google"
